Fixed unit test for job queue and moved isJobFound function as private method in job backend
