+++ HOWTO_RELEASE
@@ -1,5 +1,5 @@
 1. Test (make clean all check), fix if broken before proceeding
-2. Ensure proper version in package.json 
+2. Ensure proper version in package.json
 3. Ensure NEWS section exists for the new version, review it, add release date
 4. Drop npm-shrinkwrap.json
 5. Run npm shrinkwrap to recreate npm-shrinkwrap.json
